Compromise for unified search
Emptied out the registries, this is a gds thing and not needed for ODI.
It blows up if it tries to look for a registry as it'll expect to find one in the
government index, which won't exist. So empty hash it.

I've undone what I had previously for indexes. Tests are quite brittle and
directly tied to the idea of govuk_indexes having multiple items in there,
so don't rename it. While the naming is now a bit wrong and confusing,
by leaving it as govuk_indexes the tests will still run.

Live with it for now.