Avoid calling .blank? on nil
Previously, nil was monkeypatched to add a .blank? method (because
`lib/ses_mailer` required `aws/ses` which required `mail` which required
`mail/core_extensions/nil.rb`, which monkeypatched nil).  We've now
removed the mailer stuff, so there are a couple of places where .blank?
was being called on nil that needed fixing.